added session routes
